+++ docs/changelog.md
@@ -5,6 +5,7 @@
 - Adds support for groups on dynamic views (Show group also in dynamic view #1).
 - Fixes an issue when nested groups are configured, but groups are not nested.
 - StructurizrPlantUMLExporter: Adds support for group icons.
+- C4PlantUMLExporter: Adds support for styling groups (colours only).
 
 ## 1.12.1 (12th March 2023)